move to sensor
